+++ client/rtype/rtype/Mate.cpp
@@ -186,7 +186,7 @@ void Mate::shoot(Fire const& param)
 	uint16_t x = 0;
 	uint16_t y = 0;
 	uint8_t velocity = param.velocity;
-	uint8_t angle = param.angle;
+	float angle = param.angle;
 	//uint8_t effect = param.effect;
 
 	StaticTools::DeserializePosition(param.position, x, y);
